+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.3.5 - 2020-02-04
 
+### added 
+
+- after a field layout was removed, all labels of this layout are now deleted
+
 ### fixed
 
 - do not relabel fields in matrix blocks [#14](https://github.com/Anubarak/craft-relabel/issues/14)